Added remove button in cart page

+++ js/cart.js
@@ -43,7 +43,7 @@ if(sessionStorage.length){
         const productName = sessionStorage.key(i);
         const productObj = JSON.parse(sessionStorage.getItem(productName));
 
-        // console.log(productObj);
+        // console.log(productName);
 
         const row = document.createElement('tr');
         const imageCol = document.createElement('td');
@@ -51,6 +51,7 @@ if(sessionStorage.length){
         const priceCol = document.createElement('td');
         const quantityCol = document.createElement('td');
         const totalCol = document.createElement('td');
+        const removeCol = document.createElement('td');
 
         const imageFig = document.createElement('figure');
         const image = document.createElement('img');
@@ -58,6 +59,11 @@ if(sessionStorage.length){
         const name = document.createElement('h4');
 
         const quantityField = document.createElement('input');
+
+        const removeBtn = document.createElement('button');
+
+        removeBtn.className = 'btn btn-dark';
+        removeBtn.innerHTML = '<i class="fa fa-times"></i>';
 
         image.setAttribute('src', productObj.ProductPicUrl);
         imageFig.appendChild(image);
@@ -74,13 +80,21 @@ if(sessionStorage.length){
             calculateTotal(ev);
         };
 
+        removeBtn.onclick = (ev) => {
+            sessionStorage.removeItem(productName);
+            row.remove();
+            totalPrice.data = Number(totalPrice.data) - (Number(quantityField.value) * Number(productObj.Price));
+            purchaseMsg.textContent = '';
+        }
+
         totalCol.innerHTML = '<span>&euro;</span>' + productObj.Price;
 
         imageCol.append(imageFig);
         nameCol.append(name);
         quantityCol.appendChild(quantityField);
+        removeCol.appendChild(removeBtn);
 
-        row.append(imageCol, nameCol, priceCol, quantityCol, totalCol);
+        row.append(imageCol, nameCol, priceCol, quantityCol, totalCol, removeCol);
         tableBody.append(row);
 
         totalPrice.data = Number(totalPrice.data) + Number(productObj.Price);
@@ -89,46 +103,55 @@ if(sessionStorage.length){
 
 checkoutBtn.onclick = (ev) => {
 
-    store = db.transaction('products', 'readwrite').objectStore('products');
-    let quantityOrdered = 0;
+    if(sessionStorage.length) {
+        store = db.transaction('products', 'readwrite').objectStore('products');
+        let quantityOrdered = 0;
 
-    const rows = ev.path[3].children[0].tBodies[0].children;
+        const rows = ev.path[3].children[0].tBodies[0].children;
 
-    for(let i = 0; i < rows.length; i++) {
-        let quantityVal = rows[i].cells[3].children[0].value;
-        quantityOrdered += (+quantityVal);
-    }
+        for(let i = 0; i < rows.length; i++) {
+            let quantityVal = rows[i].cells[3].children[0].value;
+            quantityOrdered += (+quantityVal);
+        }
 
-    let date = new Date();
-    let res = store.add({totalPrice: totalPrice.data, totalQuantity: quantityOrdered, date: date.toDateString()});
+        let date = new Date();
+        let res = store.add({totalPrice: totalPrice.data, totalQuantity: quantityOrdered, date: date.toDateString()});
 
-    res.onsuccess = () => {
-        purchaseMsg.style.color = 'green';
-        purchaseMsg.style.paddingTop = '1rem';
-        purchaseMsg.textContent = 'Thanks for purchasing';
-        sessionStorage.clear();
-    }
+        res.onsuccess = () => {
+            purchaseMsg.style.color = 'green';
+            purchaseMsg.style.paddingTop = '1rem';
+            purchaseMsg.textContent = 'Thanks for purchasing';
+            sessionStorage.clear();
+            tableBody.innerHTML = '';
+        }
 
-    res.onerror = () => {
+        res.onerror = () => {
+            purchaseMsg.style.color = 'red';
+            purchaseMsg.style.paddingTop = '1rem';
+            purchaseMsg.textContent = 'Purchase was unsuccessful';
+        }
+    } else {
         purchaseMsg.style.color = 'red';
         purchaseMsg.style.paddingTop = '1rem';
-        purchaseMsg.textContent = 'Purchase was unsuccessful';
+        purchaseMsg.textContent = 'Your cart is empty';
     }
 }
 
 function calculateTotal(ev) {
-    const totalField = ev.path[2].lastElementChild.lastChild;
+    const totalField = ev.path[2].children[4].lastChild;
     const price = ev.path[2].children[2].lastChild.data;
     const quantity = ev.target.value;
     const rows = Array.from(ev.path[3].rows);
 
-    const newTotal = quantity * Number(price);
+    console.log(ev);
+
+    const newTotal = Number(quantity) * Number(price);
 
     totalField.data = newTotal;
 
     totalPrice.data = 0;
 
     rows.forEach(row => {
-        totalPrice.data = Number(totalPrice.data) + Number(row.lastChild.lastChild.data);
+        totalPrice.data = Number(totalPrice.data) + Number(row.children[4].lastChild.data);
     });
 }
